@@ -1,27 +1,28 @@
 import AppKit
-@testable import AXorcist
 import Foundation
-import XCTest
+import Testing
+@testable import AXorcist
 
-// MARK: - Action Command Tests
-
-class ActionIntegrationTests: XCTestCase {
-    // MARK: Internal
-
-    func testPerformActionSetTextEditTextAreaValue() async throws {
+@Suite(
+    "AXorcist Action Integration Tests",
+    .tags(.automation),
+    .enabled(if: AXTestEnvironment.runAutomationScenarios)
+)
+@MainActor
+struct ActionIntegrationTests {
+    @Test("Perform AXSetValue on TextEdit", .tags(.automation))
+    func performActionSetTextEditTextAreaValue() async throws {
         let actionCommandId = "performaction-setvalue-\(UUID().uuidString)"
         let queryCommandId = "query-verify-setvalue-\(UUID().uuidString)"
         let textEditBundleId = "com.apple.TextEdit"
         let textAreaRole = ApplicationServices.kAXTextAreaRole as String
         let textToSet = "Hello from AXORC performAction test! Time: \(Date())"
 
-        // Setup
         _ = try await setupTextEditAndGetInfo()
         defer { Task { await closeTextEdit() } }
 
         let textAreaLocator = Locator(criteria: [Criterion(attribute: "AXRole", value: textAreaRole)])
 
-        // Perform action
         try await performSetValueAction(
             actionCommandId: actionCommandId,
             textEditBundleId: textEditBundleId,
@@ -29,7 +30,6 @@ class ActionIntegrationTests: XCTestCase {
             textToSet: textToSet
         )
 
-        // Verify the value was set
         try await verifyTextValue(
             queryCommandId: queryCommandId,
             textEditBundleId: textEditBundleId,
@@ -38,20 +38,19 @@ class ActionIntegrationTests: XCTestCase {
         )
     }
 
-    func testExtractTextFromTextEditTextArea() async throws {
+    @Test("Extract text after setting value", .tags(.automation))
+    func extractTextFromTextEditTextArea() async throws {
         let setValueCommandId = "setvalue-for-extract-\(UUID().uuidString)"
         let extractTextCommandId = "extracttext-textedit-textarea-\(UUID().uuidString)"
         let textEditBundleId = "com.apple.TextEdit"
         let textAreaRole = ApplicationServices.kAXTextAreaRole as String
         let textToSetAndExtract = "Text to be extracted by AXORC. Unique: \(UUID().uuidString)"
 
-        // Setup
         _ = try await setupTextEditAndGetInfo()
         defer { Task { await closeTextEdit() } }
 
         let textAreaLocator = Locator(criteria: [Criterion(attribute: "AXRole", value: textAreaRole)])
 
-        // Set text value
         try await performSetValueAction(
             actionCommandId: setValueCommandId,
             textEditBundleId: textEditBundleId,
@@ -59,7 +58,6 @@ class ActionIntegrationTests: XCTestCase {
             textToSet: textToSetAndExtract
         )
 
-        // Extract and verify text
         try await extractAndVerifyText(
             extractTextCommandId: extractTextCommandId,
             textEditBundleId: textEditBundleId,
@@ -67,8 +65,6 @@ class ActionIntegrationTests: XCTestCase {
             expectedText: textToSetAndExtract
         )
     }
-
-    // MARK: Private
 
     // MARK: - Helper Functions
 
@@ -85,14 +81,14 @@ class ActionIntegrationTests: XCTestCase {
             debugLogging: true,
             locator: textAreaLocator,
             actionName: "AXSetValue",
-            actionValue: AnyCodable(textToSet)
+            actionValue: .string(textToSet)
         )
 
         let response = try await executeCommand(performActionEnvelope)
 
-        XCTAssertEqual(response.commandId, actionCommandId)
-        XCTAssertEqual(
-            response.success, true,
+        #expect(response.commandId == actionCommandId)
+        #expect(
+            response.success,
             "performAction command was not successful. Error: \(response.error?.message ?? "N/A")"
         )
 
@@ -116,9 +112,9 @@ class ActionIntegrationTests: XCTestCase {
 
         let response = try await executeCommand(queryEnvelope)
 
-        XCTAssertEqual(response.commandId, queryCommandId)
-        XCTAssertEqual(
-            response.success, true,
+        #expect(response.commandId == queryCommandId)
+        #expect(
+            response.success,
             "Query (verify) command failed. Error: \(response.error?.message ?? "N/A")"
         )
 
@@ -126,13 +122,13 @@ class ActionIntegrationTests: XCTestCase {
             throw TestError.generic("Attributes nil in query (verify) response.")
         }
 
-        let retrievedValue = attributes["AXValue"]?.value as? String
-        XCTAssertEqual(
-            retrievedValue, expectedText,
+        let retrievedValue = attributes["AXValue"]?.anyValue as? String
+        #expect(
+            retrievedValue == expectedText,
             "AXValue did not match. Expected: '\(expectedText)'. Got: '\(retrievedValue ?? "nil")'"
         )
 
-        XCTAssertNotEqual(response.debugLogs, nil)
+        #expect(response.debugLogs != nil)
     }
 
     private func extractAndVerifyText(
@@ -151,30 +147,28 @@ class ActionIntegrationTests: XCTestCase {
 
         let response = try await executeCommand(extractTextEnvelope)
 
-        XCTAssertEqual(response.commandId, extractTextCommandId)
-        XCTAssertEqual(
-            response.success, true,
+        #expect(response.commandId == extractTextCommandId)
+        #expect(
+            response.success,
             "extractText command failed. Error: \(response.error?.message ?? "N/A")"
         )
-        XCTAssertEqual(response.command, CommandType.extractText.rawValue)
+        #expect(response.command == CommandType.extractText.rawValue)
 
         guard let attributes = response.data?.attributes else {
             throw TestError.generic("Attributes nil in extractText response.")
         }
 
-        let extractedValue = attributes["AXValue"]?.value as? String
-        XCTAssertEqual(
-            extractedValue, expectedText,
+        let extractedValue = attributes["AXValue"]?.anyValue as? String
+        #expect(
+            extractedValue == expectedText,
             "Extracted text did not match. Expected: '\(expectedText)'. Got: '\(extractedValue ?? "nil")'"
         )
 
-        XCTAssertNotEqual(response.debugLogs, nil)
-        XCTAssertTrue(
-            response.debugLogs?
-                .contains { log in
-                    log.contains("Handling extractText command") ||
-                        log.contains("handleExtractText completed")
-                } == true,
+        #expect(response.debugLogs != nil)
+        #expect(
+            response.debugLogs?.contains { log in
+                log.contains("Handling extractText command") || log.contains("handleExtractText completed")
+            } == true,
             "Debug logs should indicate extractText execution."
         )
     }
@@ -191,26 +185,20 @@ class ActionIntegrationTests: XCTestCase {
         let result = try runAXORCCommand(arguments: [jsonString])
         let (output, errorOutput, exitCode) = (result.output, result.errorOutput, result.exitCode)
 
-        XCTAssertEqual(exitCode, 0, "Command failed. Error: \(errorOutput ?? "N/A")")
-        XCTAssertTrue(
-            (errorOutput == nil || errorOutput!.isEmpty),
+        #expect(exitCode == 0, "Command failed. Error: \(errorOutput ?? "N/A")")
+        #expect(
+            errorOutput?.isEmpty ?? true,
             "STDERR should be empty. Got: \(errorOutput ?? "")"
         )
 
         guard let outputString = output, !outputString.isEmpty else {
-            throw TestError.generic("Output was nil/empty.")
+            throw TestError.generic("Output string was nil or empty for command: \(command.commandId)")
         }
-
-        print("Received output: \(outputString)")
-
         guard let responseData = outputString.data(using: String.Encoding.utf8) else {
-            throw TestError.generic("Could not convert output to data.")
+            throw TestError.generic("Failed to convert output string to data for command: \(command.commandId)")
         }
 
-        do {
-            return try JSONDecoder().decode(QueryResponse.self, from: responseData)
-        } catch {
-            throw TestError.generic("Failed to decode response: \(error.localizedDescription). JSON: \(outputString)")
-        }
+        return try JSONDecoder().decode(QueryResponse.self, from: responseData)
     }
 }
+
