@@ -15,22 +15,29 @@ import Foundation
 /// - Application targeting and focus management
 @MainActor
 public extension AXorcist {
+    private func logQuery(_ level: AXLogLevel, _ parts: String...) {
+        let message = parts.joined(separator: ", ")
+        GlobalAXLogger.shared.log(AXLogEntry(level: level, message: message))
+    }
+
     // MARK: - Query Handler
 
     func handleQuery(command: QueryCommand, maxDepth externalMaxDepth: Int?) -> AXResponse {
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .info,
-            message: "HandleQuery: App '\(command.appIdentifier ?? "focused")', Locator: \(command.locator)"
-        ))
+        logQuery(
+            .info,
+            "HandleQuery: App '\(command.appIdentifier ?? "focused")'",
+            "Locator: \(command.locator)"
+        )
 
         let appIdentifier = command.appIdentifier ?? "focused"
         let resolvedMaxDepth = externalMaxDepth ?? 10
 
         // DEBUG LOG FOR MAX DEPTH
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "HandleQuery: externalMaxDepth = \(String(describing: externalMaxDepth)), resolved maxDepth = \(resolvedMaxDepth)"
-        ))
+        logQuery(
+            .debug,
+            "HandleQuery: externalMaxDepth = \(String(describing: externalMaxDepth))",
+            "resolved maxDepth = \(resolvedMaxDepth)"
+        )
 
         let (foundElement, findError) = findTargetElement(
             for: appIdentifier,
@@ -41,13 +48,13 @@ public extension AXorcist {
         guard let element = foundElement else {
             let errorMessage = findError ??
                 "HandleQuery: Element not found for app '\(appIdentifier)' with locator \(command.locator)."
-            GlobalAXLogger.shared.log(AXLogEntry(level: .error, message: errorMessage))
+            logQuery(.error, errorMessage)
             return .errorResponse(message: errorMessage, code: .elementNotFound)
         }
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "HandleQuery: Found element: \(element.briefDescription(option: ValueFormatOption.smart))"
-        ))
+        logQuery(
+            .debug,
+            "HandleQuery: Found element: \(element.briefDescription(option: ValueFormatOption.smart))"
+        )
 
         // Fetch attributes specified in command.attributesToReturn, or default if nil/empty
         let attributesToFetch = command.attributesToReturn ?? AXMiscConstants.defaultAttributesToFetch
@@ -63,11 +70,12 @@ public extension AXorcist {
     // MARK: - Get Attributes Handler
 
     func handleGetAttributes(command: GetAttributesCommand) -> AXResponse {
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .info,
-            message: "HandleGetAttrs: App '\(command.appIdentifier ?? "focused")', " +
-                "Locator: \(command.locator), Attributes: \(command.attributes.joined(separator: ", "))"
-        ))
+        logQuery(
+            .info,
+            "HandleGetAttrs: App '\(command.appIdentifier ?? "focused")'",
+            "Locator: \(command.locator)",
+            "Attributes: \(command.attributes.joined(separator: ", "))"
+        )
 
         let (foundElement, findError) = findTargetElement(
             for: command.appIdentifier ?? "focused",
@@ -76,15 +84,18 @@ public extension AXorcist {
         )
 
         guard let element = foundElement else {
-            let errorMessage = findError ??
-                "HandleGetAttrs: Element not found for app '\(command.appIdentifier ?? "focused")' with locator \(command.locator)."
-            GlobalAXLogger.shared.log(AXLogEntry(level: .error, message: errorMessage))
+            let fallbackError = [
+                "HandleGetAttrs: Element not found for app '\(command.appIdentifier ?? "focused")'",
+                "Locator: \(command.locator)"
+            ].joined(separator: ", ")
+            let errorMessage = findError ?? fallbackError
+            logQuery(.error, errorMessage)
             return .errorResponse(message: errorMessage, code: .elementNotFound)
         }
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "HandleGetAttrs: Found element: \(element.briefDescription(option: ValueFormatOption.smart))"
-        ))
+        logQuery(
+            .debug,
+            "HandleGetAttrs: Found element: \(element.briefDescription(option: ValueFormatOption.smart))"
+        )
 
         var attributesDict: [String: AXValueWrapper] = [:]
         for attrName in command.attributes {
@@ -96,18 +107,18 @@ public extension AXorcist {
         }
 
         let briefDesc = element.briefDescription(option: ValueFormatOption.smart)
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "HandleGetAttrs: Attributes for '\(briefDesc)': " +
-                "\(attributesDict.mapValues { String(describing: $0.anyValue?.value) })"
-        ))
+        logQuery(
+            .debug,
+            "HandleGetAttrs: Attributes for '\(briefDesc)'",
+            "\(attributesDict.mapValues { String(describing: $0.anyValue) })"
+        )
 
         // Log fetched attributes for debugging purposes
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "GetAttributes: Fetched attributes for \(briefDesc): " +
-                "\(attributesDict.mapValues { String(describing: $0.anyValue?.value) })"
-        ))
+        logQuery(
+            .debug,
+            "GetAttributes: Fetched attributes for \(briefDesc)",
+            "\(attributesDict.mapValues { String(describing: $0.anyValue) })"
+        )
 
         // Construct a simple payload containing just the attributes dictionary.
         // For a more structured response like AXElementData, we'd use buildQueryResponse or similar.
@@ -123,12 +134,13 @@ public extension AXorcist {
     // MARK: - Describe Element Handler
 
     func handleDescribeElement(command: DescribeElementCommand) -> AXResponse {
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .info,
-            message: "HandleDescribe: App '\(command.appIdentifier ?? "focused")', " +
-                "Locator: \(command.locator), Depth: \(command.depth), " +
-                "IncludeIgnored: \(command.includeIgnored)"
-        ))
+        logQuery(
+            .info,
+            "HandleDescribe: App '\(command.appIdentifier ?? "focused")'",
+            "Locator: \(command.locator)",
+            "Depth: \(command.depth)",
+            "IncludeIgnored: \(command.includeIgnored)"
+        )
 
         let (foundElement, findError) = findTargetElement(
             for: command.appIdentifier ?? "focused",
@@ -137,15 +149,19 @@ public extension AXorcist {
         )
 
         guard let element = foundElement else {
-            let errorMessage = findError ??
-                "HandleDescribe: Element not found for app '\(command.appIdentifier ?? "focused")' with locator \(command.locator)."
-            GlobalAXLogger.shared.log(AXLogEntry(level: .error, message: errorMessage))
+            let fallbackError = [
+                "HandleDescribe: Element not found for app '\(command.appIdentifier ?? "focused")'",
+                "Locator: \(command.locator)"
+            ].joined(separator: ", ")
+            let errorMessage = findError ?? fallbackError
+            logQuery(.error, errorMessage)
             return .errorResponse(message: errorMessage, code: .elementNotFound)
         }
-        GlobalAXLogger.shared.log(AXLogEntry(
-            level: .debug,
-            message: "HandleDescribe: Found element: \(element.briefDescription(option: ValueFormatOption.smart)). Describing tree..."
-        ))
+        logQuery(
+            .debug,
+            "HandleDescribe: Found element: \(element.briefDescription(option: ValueFormatOption.smart))",
+            "Describing tree..."
+        )
 
         let descriptionTree = describeElementTree(
             element: element,
@@ -159,9 +175,11 @@ public extension AXorcist {
 
     // MARK: - Helper Methods for Querying
 
-    internal func buildQueryResponse(element: Element, attributesToFetch: [String],
-                                     includeChildrenBrief: Bool) -> AXElementData
-    {
+    internal func buildQueryResponse(
+        element: Element,
+        attributesToFetch: [String],
+        includeChildrenBrief: Bool
+    ) -> AXElementData {
         let fetchedAttributes = fetchInstanceElementAttributes(element: element, attributeNames: attributesToFetch)
 
         // Get all possible attribute names for this element
@@ -189,9 +207,12 @@ public extension AXorcist {
         )
     }
 
-    private func describeElementTree(element: Element, depth: Int, includeIgnored: Bool,
-                                     currentDepth: Int) -> AXElementDescription
-    {
+    private func describeElementTree(
+        element: Element,
+        depth: Int,
+        includeIgnored: Bool,
+        currentDepth: Int
+    ) -> AXElementDescription {
         if !includeIgnored, element.isIgnored() {
             // Return a minimal description for an ignored element if not including them
             return AXElementDescription(
@@ -234,9 +255,10 @@ public extension AXorcist {
         )
     }
 
-    private func fetchInstanceElementAttributes(element: Element,
-                                                attributeNames: [String]) -> [String: AXValueWrapper]
-    {
+    private func fetchInstanceElementAttributes(
+        element: Element,
+        attributeNames: [String]
+    ) -> [String: AXValueWrapper] {
         var attributesDict: [String: AXValueWrapper] = [:]
         for name in attributeNames {
             if let value: Any = element.attribute(Attribute<Any>(name)) {
