@@ -1,14 +1,29 @@
-// LoggingHelpers.swift - Global logging functions for convenience and potential async offload.
-// AXorcist - Created by Sendhil Panchadsaram
+//
+//  LoggingHelpers.swift
+//  AXorcist
+//
 
-// This file previously contained @autoclosure versions of logging functions (axDebugLog, axInfoLog, etc.)
-// which wrapped calls to GlobalAXLogger.shared.log within a Task for potential async behavior.
+import Foundation
 
-// As part of a refactoring to make AXorcist fully synchronous and rely on main-thread execution
-// for all accessibility and logging operations, GlobalAXLogger was made synchronous.
-// The global logging functions (axDebugLog, axInfoLog, etc.) are now directly defined
-// as synchronous functions in GlobalAXLogger.swift, taking simple String messages.
+// MARK: - Logging Utilities
 
-// To resolve ambiguity between those synchronous String-based log functions and the
-// @autoclosure versions previously in this file, the contents of this file have been removed.
-// All logging calls should now resolve to the synchronous global functions in GlobalAXLogger.swift.
+/// Joins log message segments with comma separators for cleaner multi-part messages
+public func logSegments(_ parts: String...) -> String {
+    parts.joined(separator: ", ")
+}
+
+/// Joins log message segments from an array with comma separators
+public func logSegments(_ parts: [String]) -> String {
+    parts.joined(separator: ", ")
+}
+
+/// Formats log message segments with pipe separators (alternative style)
+public func formatLogSegments(_ parts: String...) -> String {
+    parts.joined(separator: " | ")
+}
+
+/// Describes a PID for logging (nil becomes "system")
+public func describePid(_ pid: pid_t?) -> String {
+    let pidValue = pid.map(String.init) ?? "system"
+    return "PID \(pidValue)"
+}
