@@ -7,69 +7,69 @@ import Foundation
 // MARK: - Attribute Collection Builders
 
 @MainActor
-func addBasicAttributes(to attributes: inout [String: AnyCodable], element: Element) async {
+func addBasicAttributes(to attributes: inout [String: AttributeValue], element: Element) async {
     if let role = element.role() {
-        attributes[AXAttributeNames.kAXRoleAttribute] = AnyCodable(role)
+        attributes[AXAttributeNames.kAXRoleAttribute] = .string(role)
     }
     if let subrole = element.subrole() {
-        attributes[AXAttributeNames.kAXSubroleAttribute] = AnyCodable(subrole)
+        attributes[AXAttributeNames.kAXSubroleAttribute] = .string(subrole)
     }
     if let title = element.title() {
-        attributes[AXAttributeNames.kAXTitleAttribute] = AnyCodable(title)
+        attributes[AXAttributeNames.kAXTitleAttribute] = .string(title)
     }
     if let descriptionText = element.descriptionText() {
-        attributes[AXAttributeNames.kAXDescriptionAttribute] = AnyCodable(descriptionText)
+        attributes[AXAttributeNames.kAXDescriptionAttribute] = .string(descriptionText)
     }
     if let value = element.value() {
-        attributes[AXAttributeNames.kAXValueAttribute] = AnyCodable(value)
+        attributes[AXAttributeNames.kAXValueAttribute] = AttributeValue(from: value)
     }
     if let help = element.attribute(Attribute<String>(AXAttributeNames.kAXHelpAttribute)) {
-        attributes[AXAttributeNames.kAXHelpAttribute] = AnyCodable(help)
+        attributes[AXAttributeNames.kAXHelpAttribute] = .string(help)
     }
     if let placeholder = element.attribute(Attribute<String>(AXAttributeNames.kAXPlaceholderValueAttribute)) {
-        attributes[AXAttributeNames.kAXPlaceholderValueAttribute] = AnyCodable(placeholder)
+        attributes[AXAttributeNames.kAXPlaceholderValueAttribute] = .string(placeholder)
     }
 }
 
 @MainActor
-func addStateAttributes(to attributes: inout [String: AnyCodable], element: Element) async {
-    attributes[AXAttributeNames.kAXEnabledAttribute] = AnyCodable(element.isEnabled())
-    attributes[AXAttributeNames.kAXFocusedAttribute] = AnyCodable(element.isFocused())
-    attributes[AXAttributeNames.kAXHiddenAttribute] = AnyCodable(element.isHidden())
-    attributes[AXMiscConstants.isIgnoredAttributeKey] = AnyCodable(element.isIgnored())
-    attributes[AXAttributeNames.kAXElementBusyAttribute] = AnyCodable(element.isElementBusy())
+func addStateAttributes(to attributes: inout [String: AttributeValue], element: Element) async {
+    attributes[AXAttributeNames.kAXEnabledAttribute] = .bool(element.isEnabled() ?? false)
+    attributes[AXAttributeNames.kAXFocusedAttribute] = .bool(element.isFocused() ?? false)
+    attributes[AXAttributeNames.kAXHiddenAttribute] = .bool(element.isHidden() ?? false)
+    attributes[AXMiscConstants.isIgnoredAttributeKey] = .bool(element.isIgnored())
+    attributes[AXAttributeNames.kAXElementBusyAttribute] = .bool(element.isElementBusy() ?? false)
 }
 
 @MainActor
-func addGeometryAttributes(to attributes: inout [String: AnyCodable], element: Element) async {
+func addGeometryAttributes(to attributes: inout [String: AttributeValue], element: Element) async {
     if let position = element.attribute(Attribute<CGPoint>(AXAttributeNames.kAXPositionAttribute)) {
-        attributes[AXAttributeNames.kAXPositionAttribute] = AnyCodable(NSPointToDictionary(position))
+        attributes[AXAttributeNames.kAXPositionAttribute] = AttributeValue(from: NSPointToDictionary(position))
     }
     if let size = element.attribute(Attribute<CGSize>(AXAttributeNames.kAXSizeAttribute)) {
-        attributes[AXAttributeNames.kAXSizeAttribute] = AnyCodable(NSSizeToDictionary(size))
+        attributes[AXAttributeNames.kAXSizeAttribute] = AttributeValue(from: NSSizeToDictionary(size))
     }
 }
 
 @MainActor
 func addHierarchyAttributes(
-    to attributes: inout [String: AnyCodable],
+    to attributes: inout [String: AttributeValue],
     element: Element,
     valueFormatOption _: ValueFormatOption
 ) async {
     if let parent = element.parent() {
-        attributes[AXAttributeNames.kAXParentAttribute] = AnyCodable(
+        attributes[AXAttributeNames.kAXParentAttribute] = .string(
             parent.briefDescription(option: .raw)
         )
     }
     if let children = element.children() {
-        attributes[AXAttributeNames.kAXChildrenAttribute] = AnyCodable(
-            children.map { $0.briefDescription(option: .raw) }
+        attributes[AXAttributeNames.kAXChildrenAttribute] = .array(
+            children.map { .string($0.briefDescription(option: .raw)) }
         )
     }
 }
 
 @MainActor
-func addActionAttributes(to attributes: inout [String: AnyCodable], element: Element) async {
+func addActionAttributes(to attributes: inout [String: AttributeValue], element: Element) async {
     var actionsToStore: [String]?
 
     if let currentActions = element.supportedActions(), !currentActions.isEmpty {
@@ -85,16 +85,16 @@ func addActionAttributes(to attributes: inout [String: AnyCodable], element: Ele
     }
 
     attributes[AXAttributeNames.kAXActionsAttribute] = actionsToStore != nil
-        ? AnyCodable(actionsToStore)
-        : AnyCodable(nil as [String]?)
+        ? .array(actionsToStore!.map { .string($0) })
+        : .null
 
     if element.isActionSupported(AXActionNames.kAXPressAction) {
-        attributes["\(AXActionNames.kAXPressAction)_Supported"] = AnyCodable(true)
+        attributes["\(AXActionNames.kAXPressAction)_Supported"] = .bool(true)
     }
 }
 
 @MainActor
-func addStandardStringAttributes(to attributes: inout [String: AnyCodable], element: Element) async {
+func addStandardStringAttributes(to attributes: inout [String: AttributeValue], element: Element) async {
     let standardAttributes = [
         AXAttributeNames.kAXRoleDescriptionAttribute,
         AXAttributeNames.kAXValueDescriptionAttribute,
@@ -105,13 +105,13 @@ func addStandardStringAttributes(to attributes: inout [String: AnyCodable], elem
         if attributes[attrName] == nil,
            let attrValue: String = element.attribute(Attribute<String>(attrName))
         {
-            attributes[attrName] = AnyCodable(attrValue)
+            attributes[attrName] = .string(attrValue)
         }
     }
 }
 
 @MainActor
-func addStoredAttributes(to attributes: inout [String: AnyCodable], element: Element) {
+func addStoredAttributes(to attributes: inout [String: AttributeValue], element: Element) {
     guard let stored = element.attributes else { return }
 
     for (key, val) in stored where attributes[key] == nil {
@@ -120,22 +120,22 @@ func addStoredAttributes(to attributes: inout [String: AnyCodable], element: Ele
 }
 
 @MainActor
-func addComputedProperties(to attributes: inout [String: AnyCodable], element: Element) async {
+func addComputedProperties(to attributes: inout [String: AttributeValue], element: Element) async {
     if attributes[AXMiscConstants.computedNameAttributeKey] == nil,
        let name = element.computedName()
     {
-        attributes[AXMiscConstants.computedNameAttributeKey] = AnyCodable(name)
+        attributes[AXMiscConstants.computedNameAttributeKey] = .string(name)
     }
 
     if attributes[AXMiscConstants.computedPathAttributeKey] == nil {
-        attributes[AXMiscConstants.computedPathAttributeKey] = AnyCodable(element.generatePathString())
+        attributes[AXMiscConstants.computedPathAttributeKey] = .string(element.generatePathString())
     }
 
     if attributes[AXMiscConstants.isClickableAttributeKey] == nil {
         let isButton = element.role() == AXRoleNames.kAXButtonRole
         let hasPressAction = element.isActionSupported(AXActionNames.kAXPressAction)
         if isButton || hasPressAction {
-            attributes[AXMiscConstants.isClickableAttributeKey] = AnyCodable(true)
+            attributes[AXMiscConstants.isClickableAttributeKey] = .bool(true)
         }
     }
 }
