@@ -20,10 +20,14 @@ func matchRoleAttribute(
         ))
     }
     return compareStrings(
-        actual, expectedValue, matchType,
+        actual,
+        expectedValue,
+        matchType,
         caseSensitive: false,
-        attributeName: AXAttributeNames.kAXRoleAttribute,
-        elementDescriptionForLog: elementDescriptionForLog
+        context: StringComparisonContext(
+            attributeName: AXAttributeNames.kAXRoleAttribute,
+            elementDescription: elementDescriptionForLog
+        )
     )
 }
 
@@ -37,10 +41,14 @@ func matchSubroleAttribute(
     let actual = element.subrole()
     GlobalAXLogger.shared.log(AXLogEntry(level: .debug, message: "SC/MSC/Subrole: Actual='\(actual ?? "nil")'"))
     return compareStrings(
-        actual, expectedValue, matchType,
+        actual,
+        expectedValue,
+        matchType,
         caseSensitive: false,
-        attributeName: AXAttributeNames.kAXSubroleAttribute,
-        elementDescriptionForLog: elementDescriptionForLog
+        context: StringComparisonContext(
+            attributeName: AXAttributeNames.kAXSubroleAttribute,
+            elementDescription: elementDescriptionForLog
+        )
     )
 }
 
@@ -54,10 +62,14 @@ func matchIdentifierAttribute(
     let actual = element.identifier()
     GlobalAXLogger.shared.log(AXLogEntry(level: .debug, message: "SC/MSC/ID: Actual='\(actual ?? "nil")'"))
     return compareStrings(
-        actual, expectedValue, matchType,
+        actual,
+        expectedValue,
+        matchType,
         caseSensitive: true,
-        attributeName: AXAttributeNames.kAXIdentifierAttribute,
-        elementDescriptionForLog: elementDescriptionForLog
+        context: StringComparisonContext(
+            attributeName: AXAttributeNames.kAXIdentifierAttribute,
+            elementDescription: elementDescriptionForLog
+        )
     )
 }
 
@@ -132,16 +144,24 @@ func matchComputedNameAttributes(
         if let value = element.value() as? String {
             let combinedName = (computedName ?? "") + " " + value
             return compareStrings(
-                combinedName, expectedValue, matchType,
-                attributeName: attributeName,
-                elementDescriptionForLog: elementDescriptionForLog
+                combinedName,
+                expectedValue,
+                matchType,
+                context: StringComparisonContext(
+                    attributeName: attributeName,
+                    elementDescription: elementDescriptionForLog
+                )
             )
         }
     }
 
     return compareStrings(
-        computedName, expectedValue, matchType,
-        attributeName: attributeName,
-        elementDescriptionForLog: elementDescriptionForLog
+        computedName,
+        expectedValue,
+        matchType,
+        context: StringComparisonContext(
+            attributeName: attributeName,
+            elementDescription: elementDescriptionForLog
+        )
     )
 }
