@@ -10,62 +10,117 @@ public func compareStrings(
     _ expectedValue: String,
     _ matchType: JSONPathHintComponent.MatchType,
     caseSensitive: Bool = true,
-    attributeName: String,
-    elementDescriptionForLog: String
+    context: StringComparisonContext
 ) -> Bool {
-    guard let actualValue = actualValueOptional, !actualValue.isEmpty else {
-        let isEmptyMatch = expectedValue.isEmpty && matchType == .exact
-
-        if isEmptyMatch {
-            GlobalAXLogger.shared.log(AXLogEntry(
-                level: .debug,
-                message: "SC/Compare: '\(attributeName)' on \(elementDescriptionForLog): " +
-                    "Actual is nil/empty, Expected is empty. MATCHED with type '\(matchType.rawValue)'."
-            ))
-            return true
-        } else {
-            GlobalAXLogger.shared.log(AXLogEntry(
-                level: .debug,
-                message: "SC/Compare: Attribute '\(attributeName)' on \(elementDescriptionForLog) " +
-                    "(actual: nil/empty, expected: '\(expectedValue)', type: \(matchType.rawValue)) -> MISMATCH"
-            ))
-            return false
-        }
+    if let decision = handleEmptyActualValue(
+        actualValue: actualValueOptional,
+        expectedValue: expectedValue,
+        matchType: matchType,
+        context: context
+    ) {
+        return decision
     }
 
-    let finalActual = caseSensitive ? actualValue : actualValue.lowercased()
-    let finalExpected = caseSensitive ? expectedValue : expectedValue.lowercased()
-    var result = false
+    let finalActual = formatValue(actualValueOptional!, caseSensitive: caseSensitive)
+    let finalExpected = formatValue(expectedValue, caseSensitive: caseSensitive)
+    let result = evaluateMatch(
+        finalActual: finalActual,
+        finalExpected: finalExpected,
+        matchType: matchType
+    )
 
+    let metadata = MatchResultMetadata(
+        actualValue: actualValueOptional!,
+        expectedValue: expectedValue,
+        matchType: matchType,
+        caseSensitive: caseSensitive,
+        didMatch: result
+    )
+    logMatchResult(context: context, metadata: metadata)
+    return result
+}
+
+@MainActor
+private func handleEmptyActualValue(
+    actualValue: String?,
+    expectedValue: String,
+    matchType: JSONPathHintComponent.MatchType,
+    context: StringComparisonContext
+) -> Bool? {
+    guard let actualValue, !actualValue.isEmpty else {
+        let isEmptyMatch = expectedValue.isEmpty && matchType == .exact
+        let message: String
+        if isEmptyMatch {
+            message = "SC/Compare: '\(context.attributeName)' on \(context.elementDescription): " +
+                "Actual is nil/empty, Expected is empty. MATCHED with type '\(matchType.rawValue)'."
+            GlobalAXLogger.shared.log(AXLogEntry(level: .debug, message: message))
+            return true
+        }
+        message = "SC/Compare: Attribute '\(context.attributeName)' on \(context.elementDescription) " +
+            "(actual: nil/empty, expected: '\(expectedValue)', type: \(matchType.rawValue)) -> MISMATCH"
+        GlobalAXLogger.shared.log(AXLogEntry(level: .debug, message: message))
+        return false
+    }
+    return nil
+}
+
+private func formatValue(_ value: String, caseSensitive: Bool) -> String {
+    caseSensitive ? value : value.lowercased()
+}
+
+private func evaluateMatch(
+    finalActual: String,
+    finalExpected: String,
+    matchType: JSONPathHintComponent.MatchType
+) -> Bool {
     switch matchType {
     case .exact:
-        result = (finalActual.localizedCompare(finalExpected) == .orderedSame)
+        return finalActual.localizedCompare(finalExpected) == .orderedSame
     case .contains:
-        result = finalActual.contains(finalExpected)
+        return finalActual.contains(finalExpected)
     case .regex:
-        result = (finalActual.range(of: finalExpected, options: .regularExpression) != nil)
+        return finalActual.range(of: finalExpected, options: .regularExpression) != nil
     case .prefix:
-        result = finalActual.hasPrefix(finalExpected)
+        return finalActual.hasPrefix(finalExpected)
     case .suffix:
-        result = finalActual.hasSuffix(finalExpected)
+        return finalActual.hasSuffix(finalExpected)
     case .containsAny:
-        let expectedSubstrings = finalExpected.split(separator: ",")
-            .map { String($0.trimmingCharacters(in: .whitespacesAndNewlines)) }
-        if expectedSubstrings.isEmpty, finalActual.isEmpty {
-            result = true
-        } else {
-            result = expectedSubstrings.contains { substring in
-                finalActual.contains(substring)
-            }
-        }
+        return evaluateContainsAnyMatch(actual: finalActual, expected: finalExpected)
     }
+}
 
-    let matchStatus = result ? "MATCH" : "MISMATCH"
+private func evaluateContainsAnyMatch(actual: String, expected: String) -> Bool {
+    let expectedSubstrings = expected.split(separator: ",")
+        .map { String($0.trimmingCharacters(in: .whitespacesAndNewlines)) }
+    if expectedSubstrings.isEmpty {
+        return actual.isEmpty
+    }
+    return expectedSubstrings.contains { substring in actual.contains(substring) }
+}
+
+@MainActor
+private func logMatchResult(
+    context: StringComparisonContext,
+    metadata: MatchResultMetadata
+) {
+    let matchStatus = metadata.didMatch ? "MATCH" : "MISMATCH"
     GlobalAXLogger.shared.log(AXLogEntry(
         level: .debug,
-        message: "SC/Compare: Attribute '\(attributeName)' on \(elementDescriptionForLog) " +
-            "(actual: '\(actualValue)', expected: '\(expectedValue)', type: \(matchType.rawValue), " +
-            "caseSensitive: \(caseSensitive)) -> \(matchStatus)"
+        message: "SC/Compare: Attribute '\(context.attributeName)' on \(context.elementDescription) " +
+            "(actual: '\(metadata.actualValue)', expected: '\(metadata.expectedValue)', " +
+            "type: \(metadata.matchType.rawValue), caseSensitive: \(metadata.caseSensitive)) -> \(matchStatus)"
     ))
-    return result
+}
+
+public struct StringComparisonContext {
+    let attributeName: String
+    let elementDescription: String
+}
+
+struct MatchResultMetadata {
+    let actualValue: String
+    let expectedValue: String
+    let matchType: JSONPathHintComponent.MatchType
+    let caseSensitive: Bool
+    let didMatch: Bool
 }
