@@ -2,6 +2,38 @@ import ApplicationServices
 import CoreFoundation
 import Foundation
 
+/// Manages accessibility observers for monitoring UI element changes.
+///
+/// `AXObserverManager` provides a centralized system for managing accessibility observers
+/// that monitor changes to UI elements. It handles observer lifecycle, notification routing,
+/// and ensures proper cleanup of resources.
+///
+/// ## Overview
+///
+/// The manager:
+/// - Creates and manages AXObserver instances for monitoring UI elements
+/// - Routes notifications to appropriate callbacks
+/// - Handles observer lifecycle and cleanup
+/// - Provides thread-safe observer management
+///
+/// This is a singleton class that should be accessed via ``shared``.
+///
+/// ## Topics
+///
+/// ### Getting the Shared Instance
+///
+/// - ``shared``
+///
+/// ### Managing Observers
+///
+/// - ``addObserver(for:notification:callback:)``
+/// - ``removeObserver(for:notification:)``
+/// - ``removeAllObservers(for:)``
+///
+/// ### Types
+///
+/// - ``AXNotificationCallback``
+/// - ``ObserverError``
 @MainActor
 public class AXObserverManager {
     // MARK: Lifecycle
@@ -34,82 +66,22 @@ public class AXObserverManager {
         observerLock.lock()
         defer { observerLock.unlock() }
 
-        // Check if we already have an observer for this element
         if var observerInfo = observers[elementId] {
-            // Add the callback for this notification
-            observerInfo.callbacks[notification.rawValue as CFString] = callback
-            observers[elementId] = observerInfo
-
-            // Add the notification to the existing observer
-            let error = AXObserverAddNotification(
-                observerInfo.observer,
-                element.underlyingElement,
-                notification.rawValue as CFString,
-                nil
-            )
-            if error != .success {
-                axErrorLog("Failed to add notification: \(error)")
-                throw ObserverError.addNotificationFailed(error)
-            }
-        } else {
-            // Create a new observer
-            guard let pid = element.pid() else {
-                throw ObserverError.other("Could not get PID for element")
-            }
-
-            var observer: AXObserver?
-
-            // Create the callback function for the observer
-            let observerCallback: AXObserverCallbackWithInfo = { observer, element, notification, userInfo, _ in
-                // Since we can't pass refcon through AXObserverCreateWithInfoCallback,
-                // we need to use a different approach to get back to the manager
-                AXObserverManager.shared.handleNotification(
-                    observer: observer,
-                    element: element,
-                    notification: notification,
-                    userInfo: userInfo
-                )
-            }
-
-            let error = AXObserverCreateWithInfoCallback(
-                pid,
-                observerCallback,
-                &observer
-            )
-
-            guard error == .success, let observer else {
-                axErrorLog("Failed to create observer: \(error)")
-                throw ObserverError.couldNotCreateObserver
-            }
-
-            // Add the notification
-            let addError = AXObserverAddNotification(
-                observer,
-                element.underlyingElement,
-                notification.rawValue as CFString,
-                nil
-            )
-            if addError != .success {
-                axErrorLog("Failed to add notification: \(addError)")
-                throw ObserverError.addNotificationFailed(addError)
-            }
-
-            // Get the run loop source and add it to the main run loop
-            let runLoopSource = AXObserverGetRunLoopSource(observer)
-            CFRunLoopAddSource(CFRunLoopGetMain(), runLoopSource, .defaultMode)
-
-            // Store the observer info
-            var callbacks: [CFString: AXNotificationCallback] = [:]
-            callbacks[notification.rawValue as CFString] = callback
-            let observerInfo = ObserverInfo(
-                observer: observer,
-                runLoopSource: runLoopSource,
-                callbacks: callbacks
+            try updateExistingObserver(
+                info: &observerInfo,
+                element: element,
+                notification: notification,
+                callback: callback
             )
             observers[elementId] = observerInfo
-
-            axDebugLog("Created observer for PID \(pid) with notification \(notification.rawValue)")
+            return
         }
+
+        observers[elementId] = try createObserverInfo(
+            for: element,
+            notification: notification,
+            callback: callback
+        )
     }
 
     // Remove observer for an element and notification
@@ -182,6 +154,10 @@ public class AXObserverManager {
         var callbacks: [CFString: AXNotificationCallback] = [:]
     }
 
+    private func notificationKey(for notification: AXNotification) -> CFString {
+        notification.rawValue as CFString
+    }
+
     private var observers: [ObjectIdentifier: ObserverInfo] = [:]
     private let observerLock = NSLock()
 
@@ -207,5 +183,96 @@ public class AXObserverManager {
 
         // Call the callback
         callback(observer, element, notification, userInfo)
+    }
+}
+
+@MainActor
+private extension AXObserverManager {
+    private func updateExistingObserver(
+        info: inout ObserverInfo,
+        element: Element,
+        notification: AXNotification,
+        callback: @escaping AXNotificationCallback
+    ) throws {
+        info.callbacks[self.notificationKey(for: notification)] = callback
+
+        let error = AXObserverAddNotification(
+            info.observer,
+            element.underlyingElement,
+            notification.rawValue as CFString,
+            nil
+        )
+
+        if error != .success {
+            axErrorLog("Failed to add notification: \(error)")
+            throw ObserverError.addNotificationFailed(error)
+        }
+    }
+
+    private func createObserverInfo(
+        for element: Element,
+        notification: AXNotification,
+        callback: @escaping AXNotificationCallback
+    ) throws -> ObserverInfo {
+        let pid = try pidForElement(element)
+        let observer = try makeObserver(pid: pid)
+        try addNotification(notification, to: observer, element: element)
+
+        let runLoopSource = AXObserverGetRunLoopSource(observer)
+        CFRunLoopAddSource(CFRunLoopGetMain(), runLoopSource, .defaultMode)
+
+        var callbacks: [CFString: AXNotificationCallback] = [:]
+        callbacks[self.notificationKey(for: notification)] = callback
+
+        axDebugLog("Created observer for PID \(pid) with notification \(notification.rawValue)")
+        return ObserverInfo(observer: observer, runLoopSource: runLoopSource, callbacks: callbacks)
+    }
+
+    func pidForElement(_ element: Element) throws -> pid_t {
+        guard let pid = element.pid() else {
+            throw ObserverError.other("Could not get PID for element")
+        }
+        return pid
+    }
+
+    func makeObserver(pid: pid_t) throws -> AXObserver {
+        var observer: AXObserver?
+        let axCallback: AXObserverCallbackWithInfo = { observer, element, notification, userInfo, _ in
+            AXObserverManager.shared.handleNotification(
+                observer: observer,
+                element: element,
+                notification: notification,
+                userInfo: userInfo
+            )
+        }
+
+        let creationError = AXObserverCreateWithInfoCallback(
+            pid,
+            axCallback,
+            &observer
+        )
+
+        guard creationError == .success, let observer else {
+            axErrorLog("Failed to create observer: \(creationError)")
+            throw ObserverError.couldNotCreateObserver
+        }
+        return observer
+    }
+
+    func addNotification(
+        _ notification: AXNotification,
+        to observer: AXObserver,
+        element: Element
+    ) throws {
+        let error = AXObserverAddNotification(
+            observer,
+            element.underlyingElement,
+            notification.rawValue as CFString,
+            nil
+        )
+        if error != .success {
+            axErrorLog("Failed to add notification: \(error)")
+            throw ObserverError.addNotificationFailed(error)
+        }
     }
 }
