@@ -2,8 +2,7 @@
 
 import Foundation
 
-// TODO: Consider if this should be public or internal depending on usage across modules if this were a larger project.
-// For AXHelper, internal or public within the module is fine.
+// These utilities are marked as public to allow usage from other modules that depend on AXorcist.
 
 /// Decodes a string representation of an array into an array of strings.
 /// The input string can be JSON-style (e.g., "["item1", "item2"]")
