@@ -1,20 +1,20 @@
 // AXORCMain.swift - Main entry point for AXORC CLI
 
-@preconcurrency import ArgumentParser
-import AXorcist // For AXorcist instance
+@preconcurrency import Commander
+import AXorcist
 import CoreFoundation
 import Foundation
 
-// axorcVersion is now defined in AXORCModels.swift
-// let axorcVersion = "0.1.0-dev"
-
 @main
 struct AXORCCommand: ParsableCommand {
-    static let configuration: CommandConfiguration = CommandConfiguration(
-        commandName: "axorc",
-        // Use axorcVersion from AXORCModels.swift or a shared constant place
-        abstract: "AXORC CLI - Handles JSON commands via various input methods. Version \(axorcVersion)"
-    )
+    @preconcurrency nonisolated static var commandDescription: CommandDescription {
+        let version = MainActor.assumeIsolated { axorcVersion }
+        return CommandDescription(
+            commandName: "axorc",
+            // Use axorcVersion from AXORCModels.swift or a shared constant place
+            abstract: "AXORC CLI - Handles JSON commands via various input methods. Version \(version)"
+        )
+    }
 
     // `--debug` now enables *normal* diagnostic output. Use the new `--verbose` flag for the extremely chatty logs.
     @Flag(name: .long, help: "Enable debug logging (normal detail level). Use --verbose for maximum detail.")
@@ -42,7 +42,10 @@ struct AXORCCommand: ParsableCommand {
     var noStopFirst: Bool = false
 
     @Argument(
-        help: "Read JSON payload directly from this string argument. If other input flags (--stdin, --file, --json) are used, this argument is ignored."
+        help: logSegments(
+            "Read JSON payload directly from this string argument.",
+            "Ignored when other input flags (--stdin, --file, --json) are provided."
+        )
     )
     var directPayload: String?
 
@@ -61,54 +64,146 @@ struct AXORCCommand: ParsableCommand {
         fflush(stdout)
 
         if command.command == .observe {
-            var observerSetupSucceeded = false
-            if let resultData = resultJsonString.data(using: .utf8) {
-                do {
-                    if let jsonOutput = try JSONSerialization.jsonObject(with: resultData, options: []) as? [String: Any],
-                       let success = jsonOutput["success"] as? Bool,
-                       let status = jsonOutput["status"] as? String {
-                        axInfoLog("AXORCMain: Parsed initial response for observe: success=\(success), status=\(status)")
-                        if success && status == "observer_started" {
-                            observerSetupSucceeded = true
-                            axInfoLog("AXORCMain: Observer setup deemed SUCCEEDED for observe command.")
-                        } else {
-                            axInfoLog("AXORCMain: Observer setup deemed FAILED for observe command (success=\(success), status=\(status)).")
-                        }
-                    } else {
-                        axErrorLog("AXORCMain: Failed to parse expected fields (success, status) from observe setup JSON.")
-                    }
-                } catch {
-                    axErrorLog("AXORCMain: Could not parse result JSON from observe setup to check for success: \(error.localizedDescription)")
-                }
-            } else {
-                axErrorLog("AXORCMain: Could not convert result JSON string to data for observe setup check.")
-            }
-
-            if observerSetupSucceeded {
-                axInfoLog("AXORCMain: Observer setup successful. Process will remain alive by running current RunLoop.")
-                #if DEBUG
-                    axInfoLog("AXORCMain: DEBUG mode - entering RunLoop.current.run() for observer.")
-                    RunLoop.current.run()
-                    axInfoLog("AXORCMain: DEBUG mode - RunLoop.current.run() finished.")
-                #else
-                    fputs("{\"error\": \"The 'observe' command is intended for DEBUG builds or specific use cases. " +
-                        "In release, it sets up the observer but will not keep the process alive indefinitely by itself. " +
-                        "Exiting normally after setup.\"}\n", stderr)
-                    fflush(stderr)
-                #endif
-            } else {
-                axErrorLog("AXORCMain: Observe command setup reported failure or result was not a success status. Exiting.")
-            }
+            handleObserveCommand(resultJsonString: resultJsonString, debugCLI: debug)
         } else {
             axClearLogs()
         }
     }
 
-    @MainActor mutating func run() async throws {
-        fputs("AXORCMain.run: VERY FIRST LINE EXECUTED.\n", stderr)
-        fflush(stderr)
+    @MainActor
+    private func handleObserveCommand(resultJsonString: String, debugCLI: Bool) {
+        let observerSetupSucceeded = parseObserveSetup(resultJsonString)
+        if observerSetupSucceeded {
+            axInfoLog(
+                logSegments(
+                    "AXORCMain: Observer setup successful",
+                    "Process will remain alive by running current RunLoop"
+                )
+            )
+            #if DEBUG
+            axInfoLog("AXORCMain: DEBUG mode - entering RunLoop.current.run() for observer.")
+            RunLoop.current.run()
+            axInfoLog("AXORCMain: DEBUG mode - RunLoop.current.run() finished.")
+            #else
+            let errorPayload = [
+                "{\"error\": \"The 'observe' command is intended for DEBUG builds or specific use cases.",
+                " In release, it sets up the observer but will not keep the process alive indefinitely by itself.",
+                " Exiting normally after setup.\"}\n"
+            ].joined()
+            fputs(errorPayload, stderr)
+            fflush(stderr)
+            #endif
+        } else {
+            axErrorLog(
+                logSegments(
+                    "AXORCMain: Observe command setup reported failure or result was not a success status",
+                    "Exiting"
+                )
+            )
+        }
+    }
 
-        // Configure global logger according to flags.
+    private func parseObserveSetup(_ jsonString: String) -> Bool {
+        guard let resultData = jsonString.data(using: .utf8) else {
+            axErrorLog("AXORCMain: Could not convert result JSON string to data for observe setup check.")
+            return false
+        }
+
+        do {
+            if
+                let jsonOutput = try JSONSerialization.jsonObject(with: resultData, options: []) as?
+                    [String: Any],
+                let success = jsonOutput["success"] as? Bool,
+                let status = jsonOutput["status"] as? String
+            {
+                axInfoLog(
+                    logSegments(
+                        "AXORCMain: Parsed initial response for observe",
+                        "success=\(success)",
+                        "status=\(status)"
+                    )
+                )
+                if success && status == "observer_started" {
+                    axInfoLog("AXORCMain: Observer setup deemed SUCCEEDED for observe command.")
+                    return true
+                }
+                axInfoLog(
+                    logSegments(
+                        "AXORCMain: Observer setup deemed FAILED for observe command",
+                        "success=\(success)",
+                        "status=\(status)"
+                    )
+                )
+                return false
+            }
+            axErrorLog(
+                logSegments(
+                    "AXORCMain: Failed to parse expected fields (success, status)",
+                    "from observe setup JSON"
+                )
+            )
+            return false
+        } catch {
+            axErrorLog(
+                logSegments(
+                    "AXORCMain: Could not parse result JSON from observe setup to check for success",
+                    error.localizedDescription
+                )
+            )
+            return false
+        }
+    }
+
+    func run() throws {
+        try MainActor.assumeIsolated {
+            try runMain()
+        }
+    }
+
+    @MainActor
+    private func runMain() throws {
+        configureLogging()
+        applyGlobalFlags()
+        logDebugVersion()
+
+        let inputResult = InputHandler.parseInput(
+            stdin: stdin,
+            file: file,
+            json: json,
+            directPayload: directPayload
+        )
+        axorcInputSource = inputResult.sourceDescription
+
+        let axorcistInstance = AXorcist.shared
+
+        if handleInputError(inputResult) {
+            return
+        }
+
+        guard let jsonStringFromInput = inputResult.jsonString else {
+            handleMissingInput()
+            return
+        }
+
+        logDebug(
+            logSegments(
+                "AXORCMain Test: Received jsonStringFromInput",
+                "[\(jsonStringFromInput)]",
+                "length: \(jsonStringFromInput.count)"
+            )
+        )
+
+        try decodeAndExecute(
+            jsonString: jsonStringFromInput,
+            axorcist: axorcistInstance
+        )
+
+        if debug && commandShouldPrintLogsAtEnd() {
+            flushDebugLogs()
+        }
+    }
+
+    private func configureLogging() {
         if verbose {
             GlobalAXLogger.shared.isLoggingEnabled = true
             GlobalAXLogger.shared.detailLevel = .verbose
@@ -119,137 +214,152 @@ struct AXORCCommand: ParsableCommand {
             GlobalAXLogger.shared.isLoggingEnabled = false
             GlobalAXLogger.shared.detailLevel = .minimal
         }
+    }
 
-        // Set global brute-force / stop-first flags
+    private func applyGlobalFlags() {
         axorcScanAll = scanAll
         axorcStopAtFirstMatch = !noStopFirst
-
-        // Honour timeout override
         if let timeout = timeout {
             axorcTraversalTimeout = TimeInterval(timeout)
         }
+    }
 
-        // For clarity in stderr output
-        fputs("AXORCMain.run: AXorc version \(axorcVersion) build \(axorcBuildStamp). Detail level: \(GlobalAXLogger.shared.detailLevel).\n", stderr)
-
-        // <<< TEST LOGGING START >>>
-        axErrorLog("AXORCMain.run: TEST ERROR LOG -- SHOULD ALWAYS APPEAR IN DEBUG OUTPUT IF LOGS ARE PRINTED")
-        axDebugLog("AXORCMain.run: TEST DEBUG LOG -- SHOULD APPEAR IF CLI --debug IS ON")
-        if debug {
-            fputs("AXORCMain.run: STDERR - CLI --debug IS ON. TEST LOGGING.\n", stderr)
-        }
-        // <<< TEST LOGGING END >>>
-
-        let inputResult = InputHandler.parseInput(
-            stdin: stdin,
-            file: file,
-            json: json,
-            directPayload: directPayload
+    private func logDebugVersion() {
+        guard debug || verbose else { return }
+        let version = MainActor.assumeIsolated { axorcVersion }
+        fputs(
+            logSegments(
+                "AXORCMain.run: AXorc version \(version) build \(axorcBuildStamp)",
+                "Detail level: \(GlobalAXLogger.shared.detailLevel)."
+            ) + "\n",
+            stderr
         )
+    }
 
-        let axorcistInstance = AXorcist.shared // Use the shared instance
+    private func handleInputError(_ inputResult: InputHandler.Result) -> Bool {
+        guard let error = inputResult.error else { return false }
+        respondWithError(
+            commandId: "input_error",
+            error: error,
+            logs: debug ? axGetLogsAsStrings(format: .text) : nil
+        )
+        return true
+    }
 
-        if let error = inputResult.error {
-            let collectedLogs = debug ? axGetLogsAsStrings(format: .text) : nil
-            let errorResponse = ErrorResponse(commandId: "input_error", error: error, debugLogs: collectedLogs)
-            if let jsonData = try? JSONEncoder().encode(errorResponse), let jsonString = String(data: jsonData, encoding: .utf8) {
-                print(jsonString)
-            } else {
-                print("{\"error\": \"Failed to encode error response\"}")
-            }
-            return
-        }
+    private func handleMissingInput() {
+        respondWithError(
+            commandId: "no_input",
+            error: "No valid JSON input received",
+            logs: debug ? axGetLogsAsStrings(format: .text) : nil
+        )
+    }
 
-        guard let jsonStringFromInput = inputResult.jsonString else {
-            let collectedLogs = debug ? axGetLogsAsStrings(format: .text) : nil
-            let errorResponse = ErrorResponse(commandId: "no_input", error: "No valid JSON input received", debugLogs: collectedLogs)
-            if let jsonData = try? JSONEncoder().encode(errorResponse), let jsonStr = String(data: jsonData, encoding: .utf8) {
-                print(jsonStr)
-            } else {
-                print("{\"error\": \"Failed to encode error response\"}")
-            }
-            return
-        }
-        axDebugLog("AXORCMain Test: Received jsonStringFromInput: [\(jsonStringFromInput)] (length: \(jsonStringFromInput.count))")
-
-        if let data = jsonStringFromInput.data(using: .utf8) {
-            let decoder = JSONDecoder()
-            decoder.keyDecodingStrategy = .convertFromSnakeCase
-            do {
-                let commands = try decoder.decode([CommandEnvelope].self, from: data)
-                if let command = commands.first {
-                    axDebugLog("AXORCMain Test: Decode attempt 1: Successfully decoded [CommandEnvelope] and got first command.")
-                    processAndExecuteCommand(command: command, axorcist: axorcistInstance, debugCLI: debug)
-                } else {
-                    axDebugLog("AXORCMain Test: Decode attempt 1: Decoded [CommandEnvelope] but array was empty.")
-                    let anError = NSError(domain: "AXORCErrorDomain", code: 1001, userInfo: [NSLocalizedDescriptionKey: "Decoded empty command array from [CommandEnvelope] attempt."])
-                    throw anError
-                }
-            } catch let arrayDecodeError {
-                axDebugLog("AXORCMain Test: Decode attempt 1 (as [CommandEnvelope]) FAILED. Error: \(arrayDecodeError). Will try as single CommandEnvelope.")
-                do {
-                    let command = try decoder.decode(CommandEnvelope.self, from: data)
-                    axDebugLog("AXORCMain Test: Decode attempt 2: Successfully decoded as SINGLE CommandEnvelope.")
-                    processAndExecuteCommand(command: command, axorcist: axorcistInstance, debugCLI: debug)
-                } catch let singleDecodeError {
-                    axDebugLog("AXORCMain Test: Decode attempt 2 (as single CommandEnvelope) ALSO FAILED. Error: \(singleDecodeError). Original array decode error was: \(arrayDecodeError)")
-                    let errorResponse = ErrorResponse(
-                        commandId: "decode_error",
-                        error: "Failed to decode JSON input: \(singleDecodeError.localizedDescription)",
-                        debugLogs: debug ? axGetLogsAsStrings() : nil
-                    )
-                    if let jsonData = try? JSONEncoder().encode(errorResponse), let jsonErrorString = String(data: jsonData, encoding: .utf8) {
-                        print(jsonErrorString)
-                    } else {
-                        print("{\"error\": \"Failed to encode decode error response: \(singleDecodeError.localizedDescription)\"}")
-                    }
-                    return
-                }
-            }
+    private func respondWithError(commandId: String, error: String, logs: [String]?) {
+        let errorResponse = ErrorResponse(commandId: commandId, error: error, debugLogs: logs)
+        let encoder = JSONEncoder()
+        encoder.outputFormatting = [.sortedKeys]
+        encoder.keyEncodingStrategy = .convertToSnakeCase
+        if let jsonData = try? encoder.encode(errorResponse),
+           let jsonString = String(data: jsonData, encoding: .utf8) {
+            print(jsonString)
         } else {
+            print("{\"error\": \"Failed to encode error response\"}")
+        }
+    }
+
+    private func decodeAndExecute(jsonString: String, axorcist: AXorcist) throws {
+        let decoder = JSONDecoder()
+        decoder.keyDecodingStrategy = .convertFromSnakeCase
+        guard let data = jsonString.data(using: .utf8) else {
             axDebugLog("AXORCMain Test: Failed to convert jsonStringFromInput to data.")
-            let errorResponse = ErrorResponse(commandId: "data_conversion_error", error: "Failed to convert JSON string to data", debugLogs: debug ? axGetLogsAsStrings() : nil)
-            if let jsonData = try? JSONEncoder().encode(errorResponse), let jsonErrorString = String(data: jsonData, encoding: .utf8) {
-                print(jsonErrorString)
-            } else {
-                print("{\"error\": \"Failed to encode data conversion error response\"}")
-            }
+            respondWithError(
+                commandId: "data_conversion_error",
+                error: "Failed to convert JSON string to data",
+                logs: debug ? axGetLogsAsStrings() : nil
+            )
             return
         }
 
-        // After processing all commands or if an error occurs
-        if debug && commandShouldPrintLogsAtEnd() {
-            let logMessages = axGetLogsAsStrings(format: .text)
-            if !logMessages.isEmpty {
-                fputs("\n--- Debug Logs (axorc run end) ---\n", stderr)
-                for logMessage in logMessages {
-                    fputs(logMessage + "\n", stderr)
-                }
-                fputs("--- End Debug Logs ---\n", stderr)
-                fflush(stderr)
+        do {
+            let commands = try decoder.decode([CommandEnvelope].self, from: data)
+            if let command = commands.first {
+                processAndExecuteCommand(command: command, axorcist: axorcist, debugCLI: debug)
+                return
+            }
+            logDebug("AXORCMain Test: Decode attempt 1: Decoded [CommandEnvelope] but array was empty.")
+            throw NSError(
+                domain: "AXORCErrorDomain",
+                code: 1001,
+                userInfo: [NSLocalizedDescriptionKey: "Decoded empty command array from [CommandEnvelope] attempt."]
+            )
+        } catch let arrayDecodeError {
+            logDebug(
+                logSegments(
+                    "AXORCMain Test: Decode attempt 1 (as [CommandEnvelope]) FAILED",
+                    "Error: \(arrayDecodeError)",
+                    "Will try as single CommandEnvelope"
+                )
+            )
+            do {
+                let command = try decoder.decode(CommandEnvelope.self, from: data)
+                processAndExecuteCommand(command: command, axorcist: axorcist, debugCLI: debug)
+            } catch let singleDecodeError {
+                logDebug(
+                    logSegments(
+                        "AXORCMain Test: Decode attempt 2 (as single CommandEnvelope) ALSO FAILED",
+                        "Error: \(singleDecodeError)",
+                        "Original array decode error was: \(arrayDecodeError)"
+                    )
+                )
+                respondWithError(
+                    commandId: "decode_error",
+                    error: "Failed to decode JSON input: \(singleDecodeError.localizedDescription)",
+                    logs: debug ? axGetLogsAsStrings() : nil
+                )
             }
         }
     }
 
+    private func flushDebugLogs() {
+        let logMessages = axGetLogsAsStrings(format: .text)
+        guard !logMessages.isEmpty else { return }
+        fputs("\n--- Debug Logs (axorc run end) ---\n", stderr)
+        logMessages.forEach { fputs($0 + "\n", stderr) }
+        fputs("--- End Debug Logs ---\n", stderr)
+        fflush(stderr)
+    }
+
+    private func logDebug(_ message: String) {
+        axDebugLog(message)
+    }
     private func commandShouldPrintLogsAtEnd() -> Bool {
-        // This is a simplified check. A more robust way would be to check
-        // the actual command type if it's available here.
-        // For now, if stdin is true or json is provided, assume it might be an observe command.
-        // This is imperfect.
-        if let jsonString = InputHandler.parseInput(stdin: stdin, file: file, json: json, directPayload: directPayload).jsonString,
-           let inputData = jsonString.data(using: .utf8) { // Corrected optional chaining and conditional binding
+        MainActor.assumeIsolated {
+            let parsedInput = InputHandler.parseInput(
+                stdin: stdin,
+                file: file,
+                json: json,
+                directPayload: directPayload
+            )
+            guard
+                let jsonString = parsedInput.jsonString,
+                let inputData = jsonString.data(using: .utf8)
+            else {
+                return true
+            }
             let decoder = JSONDecoder()
             decoder.keyDecodingStrategy = .convertFromSnakeCase
             if let commands = try? decoder.decode([CommandEnvelope].self, from: inputData),
-               commands.first?.command == .observe {
+               commands.first?.command == .observe
+            {
                 return false
             }
             if let command = try? decoder.decode(CommandEnvelope.self, from: inputData),
-               command.command == .observe {
+               command.command == .observe
+            {
                 return false
             }
+            return true
         }
-        return true
     }
 }
 
