@@ -7,36 +7,27 @@ import Foundation
 
 extension CommandType {
     func toAXCommand(commandEnvelope: CommandEnvelope) -> AXCommand? {
-        switch self {
-        case .query:
-            createQueryCommand(commandEnvelope)
-        case .performAction:
-            createPerformActionCommand(commandEnvelope)
-        case .getAttributes:
-            createGetAttributesCommand(commandEnvelope)
-        case .describeElement:
-            createDescribeElementCommand(commandEnvelope)
-        case .extractText:
-            createExtractTextCommand(commandEnvelope)
-        case .collectAll:
-            createCollectAllCommand(commandEnvelope)
-        case .batch:
-            createBatchCommand(commandEnvelope)
-        case .setFocusedValue:
-            createSetFocusedValueCommand(commandEnvelope)
-        case .getElementAtPoint:
-            createGetElementAtPointCommand(commandEnvelope)
-        case .getFocusedElement:
-            createGetFocusedElementCommand(commandEnvelope)
-        case .observe:
-            createObserveCommand(commandEnvelope)
-        case .ping, .stopObservation, .isProcessTrusted, .isAXFeatureEnabled,
-             .setNotificationHandler, .removeNotificationHandler, .getElementDescription:
-            nil
-        }
+        guard let builder = Self.commandBuilders[self] else { return nil }
+        return builder(commandEnvelope)
     }
 
-    private func createQueryCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private typealias CommandBuilder = (CommandEnvelope) -> AXCommand?
+
+    private static let commandBuilders: [CommandType: CommandBuilder] = [
+        .query: Self.createQueryCommand,
+        .performAction: Self.createPerformActionCommand,
+        .getAttributes: Self.createGetAttributesCommand,
+        .describeElement: Self.createDescribeElementCommand,
+        .extractText: Self.createExtractTextCommand,
+        .collectAll: Self.createCollectAllCommand,
+        .batch: Self.createBatchCommand,
+        .setFocusedValue: Self.createSetFocusedValueCommand,
+        .getElementAtPoint: Self.createGetElementAtPointCommand,
+        .getFocusedElement: Self.createGetFocusedElementCommand,
+        .observe: Self.createObserveCommand,
+    ]
+
+    private static func createQueryCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         let effectiveLocator = commandEnvelope.locator ?? Locator(criteria: [])
         return .query(QueryCommand(
             appIdentifier: commandEnvelope.application,
@@ -55,7 +46,7 @@ extension CommandType {
         ))
     }
 
-    private func createPerformActionCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
+    private static func createPerformActionCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
         guard let actionName = commandEnvelope.actionName else { return nil }
         return .performAction(PerformActionCommand(
             appIdentifier: commandEnvelope.application,
@@ -66,7 +57,7 @@ extension CommandType {
         ))
     }
 
-    private func createGetAttributesCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private static func createGetAttributesCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         .getAttributes(GetAttributesCommand(
             appIdentifier: commandEnvelope.application,
             locator: commandEnvelope.locator ?? Locator(criteria: []),
@@ -75,7 +66,7 @@ extension CommandType {
         ))
     }
 
-    private func createDescribeElementCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private static func createDescribeElementCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         .describeElement(DescribeElementCommand(
             appIdentifier: commandEnvelope.application,
             locator: commandEnvelope.locator ?? Locator(criteria: []),
@@ -85,7 +76,7 @@ extension CommandType {
         ))
     }
 
-    private func createExtractTextCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private static func createExtractTextCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         .extractText(ExtractTextCommand(
             appIdentifier: commandEnvelope.application,
             locator: commandEnvelope.locator ?? Locator(criteria: []),
@@ -95,7 +86,7 @@ extension CommandType {
         ))
     }
 
-    private func createCollectAllCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private static func createCollectAllCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         .collectAll(CollectAllCommand(
             appIdentifier: commandEnvelope.application,
             attributesToReturn: commandEnvelope.attributes,
@@ -105,29 +96,30 @@ extension CommandType {
         ))
     }
 
-    private func createBatchCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
+    private static func createBatchCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
         guard let batchSubCommands = commandEnvelope.subCommands else {
             axErrorLog("toAXCommand: Batch command missing subCommands in CommandEnvelope.")
             return nil
         }
         let axSubCommands = batchSubCommands.compactMap { subCmdEnv -> AXBatchCommand.SubCommandEnvelope? in
             guard let axSubCmd = subCmdEnv.command.toAXCommand(commandEnvelope: subCmdEnv) else {
-                axErrorLog(
-                    "toAXCommand: Failed to convert subCommand '\(subCmdEnv.commandId)' of type '\(subCmdEnv.command.rawValue)' to AXCommand."
-                )
+                let message = "Failed to convert subCommand '\(subCmdEnv.commandId)' of type " +
+                    "'\(subCmdEnv.command.rawValue)' to AXCommand."
+                axErrorLog("toAXCommand: \(message)")
                 return nil
             }
             return AXBatchCommand.SubCommandEnvelope(commandID: subCmdEnv.commandId, command: axSubCmd)
         }
         if axSubCommands.count != batchSubCommands.count {
             axErrorLog(
-                "toAXCommand: Some subCommands in batch failed to convert. Original: \(batchSubCommands.count), Converted: \(axSubCommands.count)"
+                "toAXCommand: Some subCommands in batch failed. Original: \(batchSubCommands.count), " +
+                    "Converted: \(axSubCommands.count)"
             )
         }
         return .batch(AXBatchCommand(commands: axSubCommands))
     }
 
-    private func createSetFocusedValueCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
+    private static func createSetFocusedValueCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
         guard let value = commandEnvelope.actionValue?.value as? String else {
             axErrorLog("toAXCommand: SetFocusedValue missing string value in actionValue or wrong type.")
             return nil
@@ -140,7 +132,7 @@ extension CommandType {
         ))
     }
 
-    private func createGetElementAtPointCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
+    private static func createGetElementAtPointCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
         guard let point = commandEnvelope.point else {
             axErrorLog("toAXCommand: GetElementAtPoint missing point.")
             return nil
@@ -154,7 +146,7 @@ extension CommandType {
         ))
     }
 
-    private func createGetFocusedElementCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
+    private static func createGetFocusedElementCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand {
         .getFocusedElement(GetFocusedElementCommand(
             appIdentifier: commandEnvelope.application,
             attributesToReturn: commandEnvelope.attributes,
@@ -162,17 +154,17 @@ extension CommandType {
         ))
     }
 
-    private func createObserveCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
+    private static func createObserveCommand(_ commandEnvelope: CommandEnvelope) -> AXCommand? {
         guard let notificationsList = commandEnvelope.notifications, !notificationsList.isEmpty else {
             axErrorLog("toAXCommand: Observe missing notifications list.")
             return nil
         }
-        guard let firstNotificationName = notificationsList.first,
-              let axNotification = AXNotification(rawValue: firstNotificationName)
+        guard
+            let firstNotificationName = notificationsList.first,
+            let axNotification = AXNotification(rawValue: firstNotificationName)
         else {
-            axErrorLog(
-                "toAXCommand: Invalid or unsupported notification name: \(notificationsList.first ?? "nil") for observe command."
-            )
+            let invalidName = notificationsList.first ?? "nil"
+            axErrorLog("toAXCommand: invalid notification name \(invalidName) for observe command.")
             return nil
         }
         return .observe(ObserveCommand(
